api/requirements.txt
Django==2.2.4
django-cors-headers==3.1.0
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
PyJWT==1.7.1
pytz==2019.2
